Swipe to Delete along with some Touch ups in story
I changed the name from "Select Stalkers" to "Select Guardians" and the Button "Start the stalking" to Start. I also set the minumum date possible to select in the ETAViewController in the datepicker to be the current date and time in the ViewDidLoad (may consider making a function for customization and throwing it in ViewDidLoad.